+++ Classes/Phoenix.Classes.psm1
@@ -494,6 +494,7 @@ class PhoenixBuild {
 #region 10-Core\PhoenixBackgroundOperation.ps1
 enum PhoenixBackgroundOperationState {
     Created
+    Queued
     Starting
     Running
     CancellationRequested
@@ -593,10 +594,27 @@ class PhoenixBackgroundOperation {
         $this.ErrorMessage = ''
     }
 
-    [void] MarkStarting() {
+    [void] MarkQueued() {
         $this.AssertState(
             [PhoenixBackgroundOperationState]::Created
         )
+
+        $this.State =
+            [PhoenixBackgroundOperationState]::Queued
+    }
+
+    [void] MarkStarting() {
+        if (
+            $this.State -ne
+                [PhoenixBackgroundOperationState]::Created -and
+            $this.State -ne
+                [PhoenixBackgroundOperationState]::Queued
+        ) {
+            throw (
+                "Operation '$($this.OperationId)' cannot start " +
+                "from state '$($this.State)'."
+            )
+        }
 
         $this.State =
             [PhoenixBackgroundOperationState]::Starting
@@ -660,6 +678,8 @@ class PhoenixBackgroundOperation {
     [void] MarkCancelled() {
         if (
             $this.State -ne
+                [PhoenixBackgroundOperationState]::Queued -and
+            $this.State -ne
                 [PhoenixBackgroundOperationState]::Running -and
             $this.State -ne
                 [PhoenixBackgroundOperationState]::Starting -and
@@ -722,6 +742,8 @@ class PhoenixBackgroundOperation {
 
     [bool] CanCancel() {
         return (
+            $this.State -eq
+                [PhoenixBackgroundOperationState]::Queued -or
             $this.State -eq
                 [PhoenixBackgroundOperationState]::Starting -or
             $this.State -eq